Move Timer and TexField
